requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
anyio==3.7.1
astor==0.8.1
async-timeout==5.0.1
attrs==25.1.0
autocommand==2.2.2
babel==2.17.0
backports.tarfile==1.2.0
bce-python-sdk==0.9.29
bcrypt==4.2.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coco_eval==0.0.4
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.1
cuda-bindings==12.8.0
cuda-python==12.8.0
cycler==0.12.1
Cython==3.1.0a1
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
fastapi==0.115.8
faster-coco-eval==1.6.5
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
Flask==3.1.0
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
gevent==24.11.1
geventhttpclient==2.3.3
gradio==3.11.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.12.0
httpcore==0.13.7
httpx==1.0.0b0
huggingface-hub==0.29.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lapx==0.5.11.post1
lazy_loader==0.4
-e git+https://github.com/cvg/LightGlue.git@edb2b838efb2ecfe3f88097c5fad9887d95aedad#egg=lightglue
lightning-utilities==0.12.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
more-itertools==10.3.0
motmetrics==1.4.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.27.1
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
onnx==1.17.0
onnxruntime-gpu==1.18.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.15
packaging==24.2
paddle2onnx==2.0.0
paddlepaddle-gpu==2.6.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.0.0
platformdirs==4.2.2
plotly==6.0.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-rapidjson==1.20
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
rarfile==4.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shapely==2.0.7
SimpleITK==2.4.1
simsimd==6.2.1
six==1.17.0
sklearn==0.0
sniffio==1.3.1
starlette==0.45.3
stringzilla==3.11.3
supervision==0.25.1
sympy==1.13.1
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
tomli==2.0.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.6.1
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
tritonclient==2.54.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
visualdl==3.0.0b0
vtk==9.4.1
websockets==15.0
Werkzeug==3.1.3
x2paddle==1.6.0
xmltodict==0.14.2
yarl==1.18.3
zipp==3.19.2
zope.event==5.0
zope.interface==7.2
